Move block hash types to primitives
In preparation for moving the `block` module to `primitives` move the
block hash types over there.
